Adding .gitignore, deleting huge demo, python script organizing.
